move ChildcareTimeRangeNormalizer and test back to Calendar namespace
