+++ src/Model/Serializer/ValueObject/Calendar/OpeningHours/OpeningHourNormalizer.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours;
 
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\ChildcareTimeRangeNormalizer;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\OpeningHour;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
